Patch for IP pool issues:
The problem is when we have more than one api-server, when a subnet is
deleted, the other api-servers keep the subnet allocator in their
subnet_allocator dict (in vnc_db.py). So when we create a subnet with
the same parameters (net, name, cidr,...) when we attach a port to him,
the allocated api is not always in the good range, depending on which
api-server responded to the create-port request.

- "net_update_notify" method updated to delete subnet allocators from
  subnet_allocators dict (in vnc_db.py module) from all api-servers

- subnet_allocators dict modified (in vnc_db.py) to sort subnet by virtual networks
   -->  original dict contained subnet allocators, now the dict contains dicts
        with vn_fq_name as key. So when we need to delete allocator from
        subnet_allocators dict we just parse current vn's allocators.
        This avoid parsing too much elements when updating a network.

- added method to get subnet_allocators using vn_fq_name (in vnc_db.py)

- new parameter "notified" added to delete_subnet_allocator method to delete
  allocator from subnet_allocator dict and not the IndexAllocator from
  ZK (it has already been deleted from the api-serve that received the
  delete request)

Closes-Bug: #1332645
Closes-Bug: #1365824

+++ src/config/api-server/vnc_db.py
@@ -485,14 +485,26 @@ class VncZkClient(object):
                                              self._reconnect_zk)
     # end
 
+    @staticmethod
+    def subnet_to_vn(subnet_fq_name):
+        # extract vn_fq_name form subnet_fq_name
+        return ':'.join(subnet_fq_name.split(':', 3)[:-1])
+    # end subnet_to_vn
+
     def create_subnet_allocator(self, subnet, subnet_alloc_list,
                                 addr_from_start, should_persist,
                                 start_subnet, size, alloc_unit):
         # TODO handle subnet resizing change, ignore for now
-        if subnet not in self._subnet_allocators:
+
+        vn_fq_name = self.subnet_to_vn(subnet)
+        # init _subnet_allocators
+        if vn_fq_name not in self._subnet_allocators:
+            self._subnet_allocators[vn_fq_name] = {}
+
+        if subnet not in self._subnet_allocators[vn_fq_name]:
             if addr_from_start is None:
                 addr_from_start = False
-            self._subnet_allocators[subnet] = IndexAllocator(
+            self._subnet_allocators[vn_fq_name][subnet] = IndexAllocator(
                 self._zk_client, self._subnet_path+'/'+subnet+'/',
                 size=size/alloc_unit, start_idx=start_subnet/alloc_unit,
                 reverse=not addr_from_start,
@@ -501,14 +513,24 @@ class VncZkClient(object):
                 max_alloc=self._MAX_SUBNET_ADDR_ALLOC/alloc_unit)
     # end create_subnet_allocator
 
-    def delete_subnet_allocator(self, subnet):
-        self._subnet_allocators.pop(subnet, None)
-        IndexAllocator.delete_all(self._zk_client,
-                                  self._subnet_path+'/'+subnet+'/')
+    def delete_subnet_allocator(self, subnet_name, notified=True):
+        vn_fq_name = self.subnet_to_vn(subnet_name)
+
+        if notified:
+            self._subnet_allocators[vn_fq_name].pop(subnet_name, None)
+        else:
+            IndexAllocator.delete_all(self._zk_client,
+                                   self._subnet_path+'/'+subnet_name+'/')
     # end delete_subnet_allocator
 
+    def _get_vn_subnet_allocators(self, vn_fq_name):
+        return self._subnet_allocators[vn_fq_name]
+    # end _get_vn_subnet_allocators
+
     def _get_subnet_allocator(self, subnet):
-        return self._subnet_allocators.get(subnet)
+        vn_fq_name = self.subnet_to_vn(subnet)
+
+        return self._subnet_allocators[vn_fq_name].get(subnet)
     # end _get_subnet_allocator
 
     def subnet_is_addr_allocated(self, subnet, addr):
@@ -977,7 +999,7 @@ class VncDbClient(object):
 
         if obj_dict is None:
             try:
-                (ok, obj_dict) = self._db_client_mgr.dbe_read(type, uuid)
+                (ok, obj_dict) = self.dbe_read(type, uuid)
                 if not ok:
                     return
             except NoIdError:
@@ -1310,9 +1332,13 @@ class VncDbClient(object):
                                should_persist, start_subnet, size, alloc_unit)
     # end subnet_create_allocator
 
-    def subnet_delete_allocator(self, subnet):
-        return self._zk_db.delete_subnet_allocator(subnet)
+    def subnet_delete_allocator(self, subnet, notified=True):
+        return self._zk_db.delete_subnet_allocator(subnet, notified)
     # end subnet_delete_allocator
+
+    def vn_get_allocators(self, vn_fq_name):
+        return self._zk_db._get_vn_subnet_allocators(vn_fq_name)
+    # end vn_get_allocators
 
     def uuid_vnlist(self):
         return self._object_db.uuid_vnlist()
